different access level

+++ handlers/CategoriesHandler/DeleteCategory.js
@@ -4,6 +4,10 @@ const Responses = require('../../API_Responses');
 module.exports.deleteCategory = async (event) => {
 
     try{
+        const userRole = event.requestContext.authorizer.claims['custom:role'];
+        if (userRole !== 'admin') {
+            return Responses._403({ message: 'Access denied. You must be an admin to perform this action.' });
+        }
         await connectToDatabase();
         const id = event.pathParameters.id;
         const category = await models.Category.findByIdAndRemove(id);
